SW: Activate ASSERT macro with RELEASE=0 builds
git-svn-id: https://svn.eduke32.com/eduke32@8349 1a8010ca-5511-0410-912e-c29ae57300e0

+++ source/sw/src/weapon.cpp
@@ -7748,7 +7748,7 @@ void TraverseBreakableWalls(short start_sect, int x, int y, int z, short ang, in
     {
         sect = sectlist[sectlistplc++];
 
-        ASSERT(sectlistplc < SIZ(sectlist));
+        ASSERT((uint16_t)sectlistplc < SIZ(sectlist));
 
         startwall = sector[sect].wallptr;
         endwall = startwall + sector[sect].wallnum;
@@ -7798,7 +7798,7 @@ void TraverseBreakableWalls(short start_sect, int x, int y, int z, short ang, in
             if (k < 0)
             {
                 sectlist[sectlistend++] = nextsector;
-                ASSERT(sectlistend < SIZ(sectlist));
+                ASSERT((uint16_t)sectlistend < SIZ(sectlist));
             }
         }
